Text Search Interface with support for Term and Proximity (AND) operations (#391)
This PR implements the search logic for Exact Phrase / Proximity AND base support.

It introduces the following layout for search operations:

Search() -> TextFetcher -> TextIterator impl

TextIterator can be one of the two:

ProximityIterator - contains at least two TextIterator - based on the query string
TermIterator
Added Integration testing and updated unit tests.

Also supports multiple words in the text iterators that the proximity iterator operates on. This is done by having each TextIterator support the keys search across all words (in lexical order) within the NextKey APIs, which needs the NextPosition to also perform the positional searches across the relevant key iterators

This PR supports Proximity AND, not ProximityOR. ProximityOR is more complex in that we can have nested proximity operations which have text and non text operations. This is planned work for the future.

Signed-off-by: Karthik Subbarao <karthikrs2021@gmail.com>